requirements.txt
tensorflow-gpu==2.4.1
tensorflow_probability
dvc[gdrive]